refactor(network-activity-plugin): put http code in respective directory
