+++ packages/network-activity-plugin/src/react-native/http/network-requests-registry.ts
