+++ packages/network-activity-plugin/src/react-native/http/xhr-interceptor.ts
