+++ packages/network-activity-plugin/src/react-native/http/xml-request.d.ts
